requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.3
django-import-export==4.3.1
sqlparse==0.5.2
tablib==3.7.0
